feat: Enable CI with GitHub Actions
This commit enables Continuous Integration (CI) for the project using GitHub Actions.

Native tests are now executed on Ubuntu, Windows, and macOS.
Android tests are run on a variety of emulators using the specified architectures.

This ensures that tests are automatically run on every push to the repository, providing early feedback on code changes and helping maintain code quality.

Additionally, this commit adds Develocity to the project to enable build scans.
Build scans are published to Gradle Enterprise when the CI environment variable is set to true.