Use release binaries by default (#5)
Not specifying `-Pleveldb.relsase` parameter to the build now will default to `true` instead of `false`.

* Set default release mode to true

Previously, the release mode was conditionally set to true based on a project property. This update changes the default behavior to ensure release mode is true if the property is not explicitly defined. This makes the build process more robust by preventing unintended defaults when the property is omitted.

* Add explicit return type to batch function

This change specifies the return type of the `batch` function as `Unit`, enhancing code clarity and consistency. Though Kotlin infers return types, explicitly defining `Unit` aligns with best practices and helps with maintaining uniform code style across the project.